update work on refactor
