+++ src/components/mdTextEditor/editor/controlpanel/ControlPanel.js
