+++ src/components/mdTextEditor/settingsFiles/languages/editor_NO.js
