+++ src/index.js
@@ -1,5 +1,6 @@
 // import 'bootstrap/dist/css/bootstrap.css';
 import "semantic-ui-css/semantic.min.css";
+import "./index.css";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
